Images Upload Fixed + Fixtures DateTime Issue
After help from matt with the date time issue from the fixtures in which we just wrapped our date time inside of a DateTime object which solved issue. we also fixed the upload features as it was in the wrong order in the controller which was causing a null issue.